requirements.txt
appdirs==1.4.4
asgiref==3.3.4
attrs==21.2.0
cached-property==1.5.2
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
colorama==0.4.4
defusedxml==0.7.1
Django==3.2.4
django-cors-headers==3.7.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
Flask==2.0.1
Flask-Cors==3.0.10
idna==2.10
isodate==0.6.0
itsdangerous==2.0.1
Jinja2==3.0.1
lxml==4.6.3
mariadb==1.0.7
MarkupSafe==2.0.1
Pillow==8.3.1
PyJWT==1.7.1
PyPDF2==1.26.0
pytz==2021.1
reportlab==3.5.68
requests==2.25.1
requests-file==1.5.1
requests-toolbelt==0.9.1
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.5
Werkzeug==2.0.1
zeep==4.0.0
